@@ -1,12 +1,10 @@
 <script lang="ts">
   import GitHub from "../assets/icons/github.svg";
-
-  const year = new Date().getFullYear();
 </script>
 
 <footer>
   <span class="made-by"
-    >Copyright @ {year} ChiefWoods
+    >Copyright @ {new Date().getFullYear()} ChiefWoods
   </span>
   <a
     class="github-link"
