@@ -1,7 +1,6 @@
-import App from './App.svelte'
+import App from "./App.svelte";
+import { mount } from "svelte";
 
-const app = new App({
-  target: document.body,
-})
+const app = mount(App, { target: document.body });
 
-export default app
+export default app;
